move test of failed cluster join according to comments by tsegismont
